Server_Python/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
paho-mqtt==1.6.1
PyJWT==2.8.0
bcrypt==4.1.1
python-dotenv==1.0.0
python-multipart==0.0.6
slowapi==0.1.9